Feat: GroupDetailFragment viewCreate

+++ app/src/main/java/com/sas/companymanagement/ui/group/Group.kt
@@ -17,5 +17,5 @@ data class Group(
 
     @PrimaryKey(autoGenerate = true)
     @ColumnInfo(name = "id")
-    var id: Int = 0,
+    var id: Long = 0L,
 )
